more correction of types
git-svn-id: https://svn.r-project.org/R/trunk@42242 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/extra.c
@@ -1467,7 +1467,7 @@ int winAccess(const char *path, int mode)
     {
         /* Now look for file security info, which is NT only */
         SECURITY_DESCRIPTOR *sdPtr = NULL;
-        unsigned long size = 0;
+        DWORD size = 0;
         GENERIC_MAPPING genMap;
         HANDLE hToken = NULL;
         DWORD desiredAccess = 0;
